pyproject.toml
[tool.poetry]
name = "python-kafka-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
